+++ GreyAnnouncer/AnnouncerManager.cs
@@ -1,28 +1,29 @@
 using System.Collections.Generic;
 using System;
+using System.Text;
 
 namespace greycsont.GreyAnnouncer;
                                                       
 public class AnnouncerManager
 {           
-    public  static float                              sharedCooldown = 0f;
-    private static Dictionary<string, IAnnouncer> _announcers    = new Dictionary<string, IAnnouncer>();
-
+    public  static float                          sharedCooldown = 0f;
+    private static Dictionary<string, IAnnouncer> m_announcers   = new Dictionary<string, IAnnouncer>();
+    public  static long                           playRequestId  = 0;
     #region Public Methods
     public static void RegisterAnnouncer(string name, IAnnouncer announcer)
     {
-        if (_announcers.ContainsKey(name))
+        if (m_announcers.ContainsKey(name))
         {
             Plugin.log.LogWarning($"Overwriting existing announcer: {name}");
         }
-        _announcers[name] = announcer;
+        m_announcers[name] = announcer;
     }
 
     public static void ReloadAllAnnouncers()
     {
-        if (_announcers.Count == 0) return;
+        if (m_announcers.Count == 0) return;
 
-        foreach (var announcer in _announcers.Values)
+        foreach (var announcer in m_announcers.Values)
         {
             TryReloadAnnouncerAudios(announcer);
         }
@@ -30,9 +31,9 @@ public class AnnouncerManager
 
     public static void UpdateAllAnnouncerPaths(string newPath)
     {
-        if (_announcers.Count == 0) return;
+        if (m_announcers.Count == 0) return;
 
-        foreach (var announcer in _announcers.Values)
+        foreach (var announcer in m_announcers.Values)
         {
             TryUpdateAnnouncerAudioPath(announcer, newPath);
         }
@@ -40,7 +41,7 @@ public class AnnouncerManager
 
     public static void ResetCooldown()
     {
-        foreach (var announcer in _announcers.Values)
+        foreach (var announcer in m_announcers.Values)
         {
             announcer.ResetCooldown();
         }
@@ -49,7 +50,7 @@ public class AnnouncerManager
 
     public static void ClearAudioClipsCache()
     {
-        foreach (var announcer in _announcers.Values)
+        foreach (var announcer in m_announcers.Values)
         {
             announcer.ClearAudioClipsCache();
         }
@@ -57,7 +58,7 @@ public class AnnouncerManager
 
     public static IAnnouncer GetAnnouncer(string name)
     {
-        return _announcers.TryGetValue(name, out var announcer) ? announcer : null;
+        return m_announcers.TryGetValue(name, out var announcer) ? announcer : null;
     }
     #endregion