+++ GreyAnnouncer/BasicUtil/ReflectionManager.cs
@@ -5,18 +5,18 @@ namespace greycsont.GreyAnnouncer;
 
 public class ReflectionManager
 {   
-    public static void LoadByReflection(string assemblyName, string methodName)
+    public static void LoadByReflection(string assemblyName, string methodName, object[] parameter = null)
     {
         try
         {
             Assembly assembly     = Assembly.GetExecutingAssembly();
             Type configuratorType = assembly.GetType(assemblyName);
             MethodInfo initialize = configuratorType.GetMethod( methodName, BindingFlags.Public | BindingFlags.Static);
-            initialize?.Invoke(null, null);
+            initialize?.Invoke(null, parameter);
         }
         catch (Exception ex)
         {
-            Plugin.log.LogError($"Failed to load optional module by : {ex}");
+            Plugin.log.LogError($"Failed to load {assemblyName}'s {methodName} by : {ex}");
         }
     }
 }