+++ docs/Changelog.md
@@ -1,6 +1,9 @@
 ### Known issue/TODO
 - Using translation from pointer to audiocategory to get clips
 - Added errortext in PluginConfigurator using ConfigHeader and attribute Archor:UpperLeft
+- In AudioAnnouncer, the longer audio loaded late and starting loading early may override the shorter one 
+- reflection access
+- remove index access in AudioAnnouncer
 
 ### v1.0.2
 - Fixed when you quit the mission in the water, the muffle effect still exists when playing audio.